fix: fetch hyperlink entities only once (#1064)
